+++ otus/app/Http/Controllers/Handbooks/HandbooksController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Handbooks;
 
 use App\Models\Handbook;
+use App\Policies\Abilities;
 use App\Services\Handbooks\HandbookService;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -14,80 +15,101 @@ class HandbooksController extends Controller {
     public function __construct(HandbookService $handbooksService) {
         $this->handbookService = $handbooksService;
     }
+
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index() {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::VIEW_ANY, Handbook::class);
+
         return \view('handbooks.list', [
             'handbooks' => $this->handbookService->searchHandbooks()
         ]);
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create() {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::CREATE, Handbook::class);
+
         return view('handbooks.create');
     }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::CREATE, Handbook::class);
+
         $this->handbookService->storeHandbook($request->all());
         return redirect(route('admin.handbooks.index'), 301);
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param \App\Models\Handbook $handbook
-     * @return \Illuminate\Http\Response
+     * @param Handbook $handbook
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Handbook $handbook) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::VIEW, $handbook);
+
         return view('handbooks.show', [
             'handbook' => $handbook
         ]);
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param \App\Models\Handbook $handbook
-     * @return \Illuminate\Http\Response
+     * @param Handbook $handbook
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Handbook $handbook) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::UPDATE, $handbook);
+
         return view('handbooks.edit', [
             'handbook' => $handbook
         ]);
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @param \App\Models\Handbook $handbook
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param Handbook $handbook
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, Handbook $handbook) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::UPDATE, $handbook);
+
         $this->handbookService->updateHandbook($handbook, $request->all());
         return redirect(route('admin.handbooks.index'), 301);
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param \App\Models\Handbook $handbook
-     * @return \Illuminate\Http\Response
+     * @param Handbook $handbook
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Handbook $handbook) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::DELETE, $handbook);
+
         $this->handbookService->destroyHandbook([$handbook->id]);
     }
 }
