+++ otus/app/Http/Controllers/Reviews/ReviewsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Reviews;
 
 use App\Models\Review;
+use App\Policies\Abilities;
 use App\Services\Materials\MaterialService;
 use App\Services\Reviews\ReviewService;
 use App\Services\Users\UserService;
@@ -22,22 +23,29 @@ class ReviewsController extends Controller {
     }
 
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index() {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::VIEW_ANY, Review::class);
+
         return \view('reviews.list', [
             'reviews' => $this->reviewService->searchReviews()
         ]);
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create() {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::CREATE, Review::class);
+
+
         return view('reviews.create', [
             'users' => $this->userService->searchUsers(),
             'materials' => $this->materialService->searchMaterials(),
@@ -45,35 +53,44 @@ class ReviewsController extends Controller {
     }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::CREATE, Review::class);
+
         $this->reviewService->storeReview($request->all());
         return redirect(route('admin.reviews.index'), 301);
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param \App\Models\Review $review
-     * @return \Illuminate\Http\Response
+     * @param Review $review
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Review $review) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::VIEW, $review);
+
         return view('reviews.show', [
             'review' => $review
         ]);
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param \App\Models\Review $review
-     * @return \Illuminate\Http\Response
+     * @param Review $review
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Review $review) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::UPDATE, $review);
+
         return view('reviews.edit', [
             'review' => $review,
             'users' => $this->userService->searchUsers(),
@@ -82,24 +99,29 @@ class ReviewsController extends Controller {
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @param \App\Models\Review $review
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param Review $review
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, Review $review) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::UPDATE, $review);
+
         $this->reviewService->updateReview($review, $request->all());
         return redirect(route('admin.reviews.index'), 301);
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param \App\Models\Review $review
-     * @return \Illuminate\Http\Response
+     * @param Review $review
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Review $review) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::DELETE, $review);
+
         $this->reviewService->destroyReview([$review->id]);
     }
 }
